Updated the project for assignment 2
There's a new element to take into consideration: Planes. 
Multiple flights can use the same plane, and so certain flights need to be delayed the plane is already in use. At the end of each flight there's a groom period for the plane where for its duration the plane cannot be assigned to a flight.